Documentating my version of modified Crowd Count
Leeyeeho's version was modified and corrected for my personal crowd count project. Project was a good success, but in the process I had to change few lines and add revisions to the code because I pulled the Shanghaitech Dataset from a different source to Leeyeeho's. I also added the converter.py so that anyone could put in their directory and change .json files that are currently written in my computer directory data. Any.py is code to help us check if a certain file is actually in the correct directory or not. I wonder what that paths.json do though. Shanghai Dataset can be found in Kaggle.